backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
datasets = "2.19.1"
langchain-pinecone = "0.1.1"
langchain-openai = "0.1.9"
langchain = "0.2.5"
langchain-core = "0.2.9"
langgraph = "0.1.1"
semantic-router = "^0.0.72"
serpapi = "0.1.5"
google-search-results = "2.4.2"
cohere = "^5.11.3"
python-dotenv = "^1.0.1"
ipython = "^8.29.0"
pygraphviz = "^1.14"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
